+++ src/polyfill/index.js
@@ -1,6 +1,10 @@
 // https://github.com/melonjs/melonJS/issues/1092
 import "core-js/proposals/global-this";
 
+// es10 string trim functions
+import "core-js/es/string/trim-left";
+import "core-js/es/string/trim-right";
+
 // "built-in" polyfills
 import "./console.js";
 import "./performance.js";