+++ src/text/bitmaptext.js
@@ -1,5 +1,4 @@
 import Color from "./../math/color.js";
-import * as stringUtil from "./../utils/string.js";
 import pool from "./../system/pooling.js";
 import loader from "./../loader/loader.js";
 import Renderable from "./../renderable/renderable.js";
@@ -254,7 +253,7 @@ class BitmapText extends Renderable {
 
         for (var i = 0; i < this._text.length; i++) {
             x = lX;
-            var string = stringUtil.trimRight(this._text[i]);
+            var string = this._text[i].trimRight();
             // adjust x pos based on alignment value
             var stringWidth = this.metrics.lineWidth(string);
             switch (this.textAlign) {